# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing WhatsApp devices and sending messages",
        "title": "WhatsApp Gateway API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/device": {
            "get": {
                "description": "Returns a list of all registered WhatsApp devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Get all devices",
                "responses": {
                    "200": {
                        "description": "List of devices with id and status",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/device/new": {
            "post": {
                "description": "Creates a new WhatsApp device and returns a QR code for authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Create new WhatsApp device",
                "responses": {
                    "200": {
                        "description": "Returns base64 encoded QR code",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/message/emoji": {
            "post": {
                "description": "Sends an emoji reaction to a specific message in a WhatsApp chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Send emoji reaction",
                "parameters": [
                    {
                        "description": "Reaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsapp.SendReactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reaction sent successfully",
                        "schema": {
                            "$ref": "#/definitions/whatsapp.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/message/file": {
            "post": {
                "description": "Sends a file/document message to a WhatsApp chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Send file message",
                "parameters": [
                    {
                        "description": "File message request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsapp.SendFileMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File sent successfully",
                        "schema": {
                            "$ref": "#/definitions/whatsapp.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/message/image": {
            "post": {
                "description": "Sends an image message to a WhatsApp chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Send image message",
                "parameters": [
                    {
                        "description": "Image message request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsapp.SendImageMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image sent successfully",
                        "schema": {
                            "$ref": "#/definitions/whatsapp.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/message/presence": {
            "post": {
                "description": "Sends a typing indicator/presence update to a WhatsApp chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Send presence update",
                "parameters": [
                    {
                        "description": "Presence request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsapp.SendPresenceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Presence sent successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/message/text": {
            "post": {
                "description": "Sends a text message to a WhatsApp chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Send text message",
                "parameters": [
                    {
                        "description": "Text message request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsapp.SendTextMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message sent successfully",
                        "schema": {
                            "$ref": "#/definitions/whatsapp.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "whatsapp.MessageResponse": {
            "type": "object",
            "properties": {
                "message_id": {
                    "type": "string"
                }
            }
        },
        "whatsapp.SendFileMessageRequest": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "chat_id": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "file_url": {
                    "type": "string"
                },
                "is_forwarded": {
                    "type": "boolean"
                },
                "reply_message_id": {
                    "type": "string"
                }
            }
        },
        "whatsapp.SendImageMessageRequest": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "chat_id": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "file_url": {
                    "type": "string"
                },
                "is_forwarded": {
                    "type": "boolean"
                },
                "reply_message_id": {
                    "type": "string"
                },
                "view_once": {
                    "type": "boolean"
                }
            }
        },
        "whatsapp.SendPresenceRequest": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "is_forwarded": {
                    "type": "boolean"
                }
            }
        },
        "whatsapp.SendReactionRequest": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "is_forwarded": {
                    "type": "boolean"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
        "whatsapp.SendTextMessageRequest": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "is_forwarded": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "reply_message_id": {
                    "type": "string"
                }
            }
        }
    }
}